python/requirements.txt
mysql-connector-python==8.0.33
Pillow==9.5.0
pytesseract==0.3.10
python-dotenv==1.0.0
